Make masking optional
Make masking of the data-container port, username and password optional
(default to true).

Masking these values means that they can't be set as job outputs and so
can't be consumed by dependent jobs.

See:
https://github.com/actions/runner/issues/1061

By making the masking optional, we allow a workaround for this use case,
allowing later jobs to use containers used by jobs earlier in a
workflow.